storage_service: check join params before sending them
If a node retries the JOIN_NODE_REQUEST, some of the parameters that it
attempts to send must not change in an unsafe way. Currently, on startup
we check that the cluster name and the snitch name didn't change, but we
don't verify the set of supported features yet.

This commit adds a check: if the node is going to retry joining the
group 0, it will first verify that no feature was removed from the
supported feature set from the previous attempt. The key
`initial_supported_features` in `system.scylla_local` is used for
persistence.