main: forbid changing the snitch on restart
This is currently checked by gossip shadow round that happens every time
the cluster starts, but we can as well persist it locally and check
whether it has changed on reboot. This reduces our reliance on gossip
(shadow round), and we can skip the check completely in topology on raft
mode as the cluster name verification is not being done in JOIN NODE
handshake.