storage_service: infer which topology operations to use on startup
Adds a piece of logic to storage_service::join_cluster which chooses the
mode in which it will boot.

If the experimental raft topology flag is disabled, it will fall back to
legacy node operations.

When the node starts for the first time, it will perform group 0
discovery. If the node creates a cluster, it will start it in raft
topology mode. If it joins an existing one, it will ask the node chosen
by the discovery algorithm about which joining method to use.

If the node is already a part of the cluster, it will base its decision
on the group0 state.